# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
      "/tue/exam/single-problem/submit": {
        "post": {
          "description": "提交单个问题并验证答案",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "tags": [
            "Exam"
          ],
          "summary": "提交单个问题并验证答案",
          "parameters": [
            {
              "description": "提交信息",
              "name": "request",
              "in": "body",
              "required": true,
              "schema": {
                "$ref": "#/definitions/exam.SubmitExamRequest"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/entity.CommonResponse-exam_SubmitProblemResponse"
              }
            }
          }
        }
      },
...
        "/tue/exam/{id}/records": {
            "get": {
                "description": "获取用户的考试评分结果",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "考试"
                ],
                "summary": "获取考试结果",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "考试记录ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.CommonResponse-schema_ExamUserRecord"
                        }
                    }
                }
            }
        },
        "/tue/exam/{id}/rescore": {
            "post": {
                "description": "管理员重新评分考试",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "考试"
                ],
                "summary": "重新评分考试",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "考试记录ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/tue/exam/{id}/submit": {
            "post": {
                "description": "提交用户的考试答案并进行评分",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "考试"
                ],
                "summary": "提交考试答案",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "考试ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "提交信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/exam.SubmitExamRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.CommonResponse-exam_SubmitExamResponse"
                        }
                    }
                }
            }
        },
        "/tue/problem/create": {
            "post": {
                "description": "创建单个题目",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Problem"
                ],
                "summary": "创建单个题目",
                "parameters": [
                    {
                        "description": "题目结构",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schema.Problem"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "返回数据",
                        "schema": {
                            "$ref": "#/definitions/entity.CommonResponse-schema_Problem"
                        }
                    }
                }
            }
        },
        "/tue/problem/list": {
            "get": {
                "description": "分页获取题目列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Problem"
                ],
                "summary": "分页获取题目列表",
                "parameters": [
                    {
                        "type": "integer",
                        "name": "end_time",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "分页参数",
                        "name": "page_num",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "sort_expr",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "start_time",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "返回数据",
                        "schema": {
                          "$ref": "#/definitions/entity.CommonResponse-entity_PaginatedTotalResponse-schema_Problem"
                        }
                    }
                }
            }
        },
      "/tue/problem/make": {
        "post": {
          "description": "创建题目",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "tags": [
            "Problem"
          ],
          "summary": "创建题目",
          "parameters": [
            {
              "description": "题目要求",
              "name": "req",
              "in": "body",
              "required": true,
              "schema": {
                "$ref": "#/definitions/course.MakeQuestionRequest"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "生成的题目",
              "schema": {
                "$ref": "#/definitions/entity.CommonResponse-schema_Problem"
                        }
                    }
                }
            }
        },
        "/tue/problem/{id}": {
            "get": {
                "description": "获取单个题目",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Problem"
                ],
                "summary": "获取单个题目",
                "parameters": [
                    {
                        "type": "string",
                        "description": "题目 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "返回数据",
                        "schema": {
                            "$ref": "#/definitions/entity.CommonResponse-schema_Problem"
                        }
                    }
                }
            }
        },
      "/tue/problem/{id}/delete": {
        "post": {
          "description": "删除题目",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "tags": [
            "Problem"
          ],
          "summary": "删除题目",
          "parameters": [
            {
              "type": "integer",
              "description": "题目 ID",
              "name": "id",
              "in": "path",
              "required": true
            }
          ],
          "responses": {
            "200": {
              "description": "删除成功与否",
              "schema": {
                "$ref": "#/definitions/entity.CommonResponse-bool"
              }
            }
          }
        }
      },
      "/tue/problem/{id}/update": {
        "post": {
          "description": "更新题目",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "tags": [
            "Problem"
          ],
          "summary": "更新题目",
          "parameters": [
            {
              "type": "integer",
              "description": "题目 ID",
              "name": "id",
              "in": "path",
              "required": true
            },
            {
              "description": "题目参数",
              "name": "problem",
              "in": "body",
              "required": true,
              "schema": {
                "$ref": "#/definitions/entity.ReqUpdateBody-schema_Problem"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "更新成功与否",
              "schema": {
                "$ref": "#/definitions/entity.CommonResponse-bool"
              }
            }
          }
        }
      },
...
    },
    "definitions": {
...
      "course.MakeQuestionRequest": {
        "type": "object",
        "required": [
          "description",
          "type"
        ],
        "properties": {
          "description": {
            "type": "string"
          },
          "type": {
            "$ref": "#/definitions/schema.ProblemType"
          }
        }
      },
...
        "entity.CommonResponse-bool": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "代码",
                    "type": "integer"
                },
                "data": {
                    "description": "数据",
                    "type": "boolean"
                },
                "msg": {
                    "description": "消息",
                    "type": "string"
                }
            }
        },
...
      "entity.CommonResponse-entity_PaginatedTotalResponse-schema_Problem": {
        "type": "object",
        "properties": {
          "code": {
            "description": "代码",
            "type": "integer"
          },
          "data": {
            "description": "数据",
            "allOf": [
              {
                "$ref": "#/definitions/entity.PaginatedTotalResponse-schema_Problem"
              }
            ]
          },
                "msg": {
                    "description": "消息",
                    "type": "string"
                }
            }
        },
...
        "entity.CommonResponse-exam_SubmitExamResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "代码",
                    "type": "integer"
                },
                "data": {
                    "description": "数据",
                    "allOf": [
                        {
                            "$ref": "#/definitions/exam.SubmitExamResponse"
                        }
                    ]
                },
              "msg": {
                "description": "消息",
                "type": "string"
              }
            }
        },
      "entity.CommonResponse-exam_SubmitProblemResponse": {
        "type": "object",
        "properties": {
          "code": {
            "description": "代码",
            "type": "integer"
          },
          "data": {
            "description": "数据",
            "allOf": [
              {
                "$ref": "#/definitions/exam.SubmitProblemResponse"
              }
            ]
          },
                "msg": {
                    "description": "消息",
                    "type": "string"
                }
            }
        },
...
        "entity.CommonResponse-schema_ExamUserRecord": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "代码",
                    "type": "integer"
                },
                "data": {
                    "description": "数据",
                    "allOf": [
                        {
                            "$ref": "#/definitions/schema.ExamUserRecord"
                        }
                    ]
                },
                "msg": {
                    "description": "消息",
                    "type": "string"
                }
            }
        },
...
        "entity.CommonResponse-schema_Problem": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "代码",
                    "type": "integer"
                },
                "data": {
                    "description": "数据",
                    "allOf": [
                        {
                            "$ref": "#/definitions/schema.Problem"
                        }
                    ]
                },
                "msg": {
                    "description": "消息",
                    "type": "string"
                }
            }
        },
...
      "entity.PaginatedTotalResponse-schema_Problem": {
        "type": "object",
        "properties": {
          "list": {
            "type": "array",
            "items": {
              "$ref": "#/definitions/schema.Problem"
            }
          },
          "total": {
            "type": "integer"
          }
        }
      },
...
      "entity.ReqUpdateBody-schema_Problem": {
        "type": "object",
        "required": [
          "data",
          "updates"
        ],
        "properties": {
          "data": {
            "$ref": "#/definitions/schema.Problem"
          },
          "updates": {
            "type": "array",
            "minItems": 1,
            "items": {
              "type": "string"
            }
          }
        }
      },
...
        "exam.SubmitExamRequest": {
            "type": "object",
            "properties": {
                "answers": {
                    "description": "答案列表",
                    "type": "array",
                  "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/exam.SubmitExamRequestAnswer"
                    }
                },
                "time_spent": {
                    "description": "答题用时（秒）",
                    "type": "integer"
                }
            }
        },
        "exam.SubmitExamRequestAnswer": {
            "type": "object",
            "properties": {
                "answer": {},
                "problem_id": {
                    "type": "integer"
                }
            }
        },
        "exam.SubmitExamResponse": {
            "type": "object",
            "properties": {
                "record_id": {
                    "description": "记录ID",
                    "type": "integer"
                }
            }
        },
      "exam.SubmitProblemResponse": {
        "type": "object",
        "properties": {
          "comment": {
            "type": "string"
          },
          "score": {
            "type": "integer"
                }
            }
        },
...
        "schema.Exam": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "description": "考试描述",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "limit_time": {
                    "description": "考试限时（单位：秒）",
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "problems": {
                    "description": "考试包含的大题",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schema.ExamProblem"
                    }
                },
                "subjects": {
                    "description": "所属科目分类",
                    "type": "string"
                },
                "total_score": {
                    "description": "考试总分（单位：0.01分）",
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "schema.ExamUserRecord": {
            "type": "object",
            "properties": {
                "answers": {
                    "description": "用户答案（关联表储存）",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schema.ExamUserRecordAnswer"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "exam": {
                    "description": "组装字段",
                    "allOf": [
                        {
                            "$ref": "#/definitions/schema.Exam"
                        }
                    ]
                },
                "exam_id": {
                    "description": "考试ID",
                    "type": "integer"
                },
                "id": {
                    "description": "普通字段",
                    "type": "integer"
                },
                "status": {
                    "description": "总评分状态",
                    "allOf": [
                        {
                            "$ref": "#/definitions/schema.ScoreStatus"
                        }
                    ]
                },
                "time_spent": {
                    "description": "答题用时（单位：秒）",
                    "type": "integer"
                },
                "total_score": {
                    "description": "总得分（单位：0.01分）",
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "description": "用户ID",
                    "type": "integer"
                }
            }
        },
        "schema.ExamUserRecordAnswer": {
            "type": "object",
            "properties": {
                "answer": {
                    "description": "用户答案"
                },
                "comments": {
                    "description": "评语/反馈",
                    "type": "string"
                },
                "exam_id": {
                    "description": "测验 ID",
                    "type": "integer"
                },
                "problem_id": {
                    "description": "题目 ID",
                    "type": "integer"
                },
                "score": {
                    "description": "得分（单位：0.01分）",
                    "type": "integer"
                },
                "status": {
                    "description": "评分状态",
                    "allOf": [
                        {
                            "$ref": "#/definitions/schema.ScoreStatus"
                        }
                    ]
                }
            }
        },
...
        "schema.Problem": {
            "type": "object",
            "properties": {
                "answer": {
                    "description": "答案（JSON存储ProblemAnswer）",
                    "allOf": [
                        {
                            "$ref": "#/definitions/schema.ProblemAnswer"
                        }
                    ]
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "description": "支持HTML/Markdown",
                    "type": "string"
                },
                "difficulty": {
                    "description": "难度等级 1-5",
                  "type": "integer",
                  "maximum": 5,
                  "minimum": 1
                },
                "explanation": {
                    "description": "答案解析",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "options": {
                    "description": "选项（JSON存储ProblemOption数组）",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schema.ProblemOption"
                    }
                },
                "subject": {
                    "description": "所属科目/分类",
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/schema.ProblemType"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "schema.ProblemAnswer": {
            "type": "object",
            "properties": {
                "answer": {
                    "description": "选择题：存储正确选项ID []uint\n填空题：存储多个填空关键词 []string\n判断题：true/false\n简答题：文本答案 string"
                }
            }
        },
        "schema.ProblemOption": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "correct": {
                    "description": "是否正确答案",
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                }
            }
        },
        "schema.ProblemType": {
            "type": "string",
            "enum": [
                "single_choice",
                "multiple_choice",
                "fill_blank",
                "short_answer",
                "true_false"
            ],
            "x-enum-varnames": [
                "SingleChoice",
                "MultipleChoice",
                "FillBlank",
                "ShortAnswer",
                "TrueFalse"
            ]
        },
...
        "schema.ScoreStatus": {
            "type": "string",
            "enum": [
                "pending",
                "scoring",
                "completed",
                "failed"
            ],
            "x-enum-comments": {
                "StatusCompleted": "评分完成",
                "StatusFailed": "评分失败",
                "StatusPending": "待评分",
                "StatusScoring": "评分中"
            },
            "x-enum-varnames": [
                "StatusPending",
                "StatusScoring",
                "StatusCompleted",
                "StatusFailed"
            ]
        },
...
    }
}